remove high-level train method (cli and python api)
The high-level training method is a _very_ thin wrapper over keras, and it does not seem useful. It would be preferable for users to write their own training functions with `tf.keras`, because then they can enjoy a powerful and flexible API. Nobrainer should, however, include a high-level inference API, both in python and command-line.